git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/branch/DPlay8@90 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/_Utility/HitTest.cpp
@@ -175,7 +175,7 @@ void    HitTest::Collide(HitTest*           pHitTest,
                 double   a = dV * dV;
                 if (a > 0.0)   //Shouldn't be a problem but ... float do have their limits of resolution
                 {
-                    tMax = (float)((halfB + (float)sqrt(halfB*halfB - a * c)) / a);
+                    tMax = (float)((halfB + (double)sqrt(halfB*halfB - a * c)) / a); // mmf cast sqrt to double instead of float
                 }
             }
